+++ CHANGES.rst
@@ -4,6 +4,7 @@ phenodata changelog
 
 development
 ===========
+- Add "--show-ids" parameter to show IDs alongside resolved text representation when using "--humanize"
 
 2018-03-14 0.5.0
 ================