+++ CHANGES.rst
@@ -5,6 +5,7 @@ phenodata changelog
 development
 ===========
 - Add "--show-ids" parameter to show IDs alongside resolved text representation when using "--humanize"
+- More compact date output for "tabular" mode
 
 2018-03-14 0.5.0
 ================