+++ CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - Add "--show-ids" parameter to show IDs alongside resolved text representation when using "--humanize"
 - More compact date output for "tabular" mode
+- Fix datetime coercion when ocourring invalid datetime values
 
 2018-03-14 0.5.0
 ================