+++ CHANGES.rst
@@ -11,6 +11,7 @@ development
 - Implement predefined sets of parameters using ``presets.json``.
   Apply with e.g. "--species-preset=mellifera-de-primary".
   See also https://community.hiveeyes.org/t/phanologischer-kalender-entwicklung/664/23.
+- Reduce logging verbosity. Add progress bar for data acquisition step.
 
 2018-03-14 0.5.0
 ================