+++ CHANGES.rst
@@ -4,6 +4,9 @@ phenodata changelog
 
 development
 ===========
+
+2018-03-14 0.6.0
+================
 - Add "--show-ids" parameter to show IDs alongside resolved text representation when using "--humanize"
 - More compact date output for "tabular" mode
 - Fix datetime coercion when encountering invalid datetime values